Handle ControlContactors and Setup frames
Add UI to reflect the "contactor setup" state which
seems to be required before the SBox does anything
with the contactors.

Process the two ContolContactors and Setup frames
and update the "Car" state immediately. Don't attempt
to validate the CRC and counter at this point.

Using the "match" statement implies requiring python
3.10 or newer.